tidycwl: Tidy Common Workflow Language Tools and Workflows

The Common Workflow Language <https://www.commonwl.org/> is an open standard for describing data analysis workflows. This package takes the raw Common Workflow Language workflows encoded in JSON or 'YAML' and turns the workflow elements into tidy data frames or lists. A graph representation for the workflow can be constructed and visualized with the parsed workflow inputs, outputs, and steps. Users can embed the visualizations in their 'Shiny' applications, and export them as HTML files or static images.

See also

Author

Maintainer: Soner Koc soner.koc@sevenbridges.com (ORCID)

Authors:

Jeffrey Grover jeffrey.grover@sevenbridges.com (ORCID)

Nan Xiao (ORCID)

Dennis Dean dennis.dean@sevenbridges.com (ORCID)

Other contributors:

Seven Bridges Genomics [copyright holder, funder]